requirements.txt
flask-cors
pillow
rdkit
decimer